Add IPTV endpoints, remove legacy server.py, fix teardown zombie and proto float bugs (Core)
- Add M3U/XMLTV pure functions (iptv.py) and wire /iptv/channels.m3u and
  /iptv/guide.xml routes into ProgramDirector (20 contract tests)
- Delete dead legacy server.py, its tests, and examples; clean up all references
- INV-TEARDOWN-AIR-REAP-001: Always call producer.stop() in _finalize_teardown
  before dropping reference, preventing orphaned AIR zombie processes (3 tests)
- Fix LINGER_SKIP: capture asyncio event loop via FastAPI startup event so
  ChannelManager linger timers work from executor threads
- INV-BLOCKPLAN-PROTO-INT-001: Coerce float ms fields to int at proto boundary;
  fix schedule_compiler source (duration_sec * 1000) and harden all JSON
  deserialization paths to prevent protobuf TypeError (5 tests)

+++ pkg/core/src/retrovue/runtime/schedule_compiler.py
@@ -806,7 +806,7 @@ def _resolve_template_segments(
             "asset_id": asset_id,
             "asset_uri": asset_meta.file_uri or "",
             "asset_start_offset_ms": 0,
-            "segment_duration_ms": asset_meta.duration_sec * 1000,
+            "segment_duration_ms": int(asset_meta.duration_sec * 1000),
             "source_type": source_type,
             "source_name": source_name,
             "is_primary": is_primary,
